Skipper pointed out that for the formula   0 + a + b + a:b we were generating too many columns for the a:b term, resulting in an overspecified model matrix. This commit fixes that.
Our strategy for avoiding overlaps is:
 1) expand each interaction out into a list of all lower-order
    interactions it could overlap with
 2) toss out all the lower-order interactions that have already been used
 3) re-combine whatever is left to produce a more concise coding scheme
The problem is that we implemented step (3) by updating some
interaction objects in-place... and these interaction objects were
also stashed away in the record of which interactions had been used
for step (2). Therefore we lost track of which interactions had
already appeared.

This commit largely rewrites the relevant code to make it much more
explicit overall, and avoid all in-place updates. Also more tests are
added to cover these cases. Also charlton.redundancy now has 100%
statement coverage, which is nice.