Make stateful transforms act more like regular functions
We used to have a special dict naming all stateful transforms that we
passed around out-of-band. Now we just look up functions in the
environment, and any function with a __charlton_stateful_transform__
attribute is a stateful transform.

Also, stateful transforms are now callable as regular functions
outside of formula context.