Better error reporting for exceptions in formula evaluation
In Python 3, use the new 'raise from' support to wrap all exceptions
generated inside EvalFactors into PatsyError's with proper origin
information.

In Python 2, we can't do that without destroying the traceback, which
would be rude. But if the internal exception was already a PatsyError,
we at least annotate it with proper origin information.

Also, improve the error message for unrecognized levels in a
CategoricalTransform (thanks to Skipper for, uh, tripping on this).

Closes #11.