Fix: Add compat code for pd.Categorical in pandas>=0.15
pandas renamed pd.Categorical.labels to pd.Categorical.codes and
pd.Categorical.levels to pd.Categorical.categories. The 'codes and
level' constructor was also removed in favour of the 'values and
categories' one, so use the Categories.from_codes(...).

It's also now possible to have Categoricals as blocks, so Series
can contain Categoricals. Unfortunately numpy.dtypes do not compare
to 'category', so use the pandas function 'is_categorical_dtype'
for this.

Added FIXMEs to all sections which should can be removed when patsy
only supports pandas >0.15.